🗑️ #90 Restructure test dirs
