+++ packages/pbt/test/Gen/Gen.Filter.test.ts
@@ -1,6 +1,6 @@
 import fc from 'fast-check';
 import * as dev from '../../src';
-import * as domainGen from './Helpers/domainGen';
+import * as domainGen from '../Helpers/domainGen';
 
 test('snapshot', () => {
   for (let i = 0; i <= 10; i++) {
