+++ packages/pbt/test/Gen/Gen.Repeat.test.ts
@@ -1,6 +1,6 @@
 import fc from 'fast-check';
 import * as dev from '../../src';
-import * as domainGen from './Helpers/domainGen';
+import * as domainGen from '../Helpers/domainGen';
 
 it('is repeatable', () => {
   fc.assert(
