+++ packages/pbt/test/Gen/Range.test.ts
@@ -1,6 +1,6 @@
 import fc from 'fast-check';
 import { Bounds, Range } from '../../src/Gen/Range';
-import * as domainGen from './Helpers/domainGen';
+import * as domainGen from '../Helpers/domainGen';
 
 type RangeParams = {
   min: number;
